Battle system (alpha) + new location
I have created battle system already. The difficulty of enemies should
be dependent on your character’s age and HP but the numbers need to be
tweaked a bit.

It works more or less but the weapon selection needs to be finished so
it’s best if you get gun in Curling street - this was tested and works
best.